API/requirements.txt
# FastAPI Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client and WebSockets
httpx==0.25.2
websockets==12.0

# Background Task Processing
celery==5.3.4
redis==5.0.1

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# File Handling and Storage
aiofiles==23.2.1
python-multipart==0.0.6

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database (Optional - for persistent storage)
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Marketing Agent Dependencies (inherit from main project)
# Note: These should match your main requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langsmith==0.0.69
langgraph==0.0.19
openai==1.6.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
